sources: refactor zip source into module.
Extract the Zip source into `sources/_zip.py`.

LP: #1648906
