+++ snapcraft/tests/sources/test_sources.py
@@ -133,40 +133,6 @@ class TestTar(tests.FakeFileHTTPServerBasedTestCase):
         self.assertTrue(os.path.exists(os.path.join('dst', 'link.txt')))
 
 
-class TestZip(tests.FakeFileHTTPServerBasedTestCase):
-
-    @unittest.mock.patch('zipfile.ZipFile')
-    def test_pull_zipfile_must_download_and_extract(self, mock_zip):
-        dest_dir = 'src'
-        os.makedirs(dest_dir)
-        zip_file_name = 'test.zip'
-        source = 'http://{}:{}/{file_name}'.format(
-            *self.server.server_address, file_name=zip_file_name)
-        zip_source = sources.Zip(source, dest_dir)
-
-        zip_source.pull()
-
-        mock_zip.assert_called_once_with(
-            os.path.join(zip_source.source_dir, zip_file_name))
-
-    @unittest.mock.patch('zipfile.ZipFile')
-    def test_extract_and_keep_zipfile(self, mock_zip):
-        zip_file_name = 'test.zip'
-        source = 'http://{}:{}/{file_name}'.format(
-            *self.server.server_address, file_name=zip_file_name)
-        dest_dir = os.path.abspath(os.curdir)
-        zip_source = sources.Zip(source, dest_dir)
-
-        zip_source.download()
-        zip_source.provision(dst=dest_dir, keep_zip=True)
-
-        zip_download = os.path.join(zip_source.source_dir, zip_file_name)
-        mock_zip.assert_called_once_with(zip_download)
-
-        with open(zip_download, 'r') as zip_file:
-            self.assertEqual('Test fake compressed file', zip_file.read())
-
-
 class TestDeb(tests.FakeFileHTTPServerBasedTestCase):
 
     def setUp(self):
